refactor: consistent file structure and naming
