+++ unicon_runner/executor/variants/podman/executor.py
@@ -1,7 +1,7 @@
 import asyncio
 import os
 
-from unicon_runner.executor.variants.base import Executor, ExecutorResult
+from unicon_runner.executor.base import Executor, ExecutorResult
 from unicon_runner.schemas import Request, Status
