+++ unicon_runner/executor/variants/sandbox/executor.py
@@ -3,7 +3,7 @@ import os
 
 from jinja2 import Environment, FileSystemLoader, select_autoescape
 
-from unicon_runner.executor.variants.base import Executor, ExecutorResult
+from unicon_runner.executor.base import Executor, ExecutorResult
 from unicon_runner.schemas import Request, Status
