+++ unicon_runner/runner/runner.py
@@ -1,7 +1,7 @@
 from collections.abc import Mapping
 from enum import Enum
 
-from unicon_runner.executor.variants.base import Executor, ExecutorResult
+from unicon_runner.executor.base import Executor, ExecutorResult
 from unicon_runner.executor.variants.podman.executor import PodmanExecutor
 from unicon_runner.executor.variants.sandbox.executor import SandboxExecutor
 from unicon_runner.executor.variants.unsafe.executor import UnsafeExecutor
